Damit fällt der Builder, der eigentlich keiner ist, mit der Countries-Klasse zusammen. Das trifft sich gut, denn wir waren uns ja eh nicht so ganz sicher, ob die Trennung zwischen der Polygon-Seite und der Regel-Reflection-Seite unbedingt genau so sein muss, wie sie ist. Also erstmal flachklopfen.
